ui/requirements.txt
flask
kafka-python
python-dateutil